backend/requirements.txt
# Latest stable versions compatible with Python 3.12
openai>=1.50.0
anthropic>=0.40.0
python-dotenv>=1.0.0
typing_extensions>=4.8.0
langchain-openai>=0.1.0
langchain-core>=0.3.0
langchain>=0.3.0
langchain-anthropic>=0.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Optional: Storage backends
# Uncomment the line below if you want to use Supabase storage
supabase>=2.0.0